# main_servise/openapi.yml
openapi: 3.1.0
info:
  title: Main Service
  version: 0.2.0
servers:
  - url: https://localhost:5005
paths:
  "/register":
    post:
      tags:
      - login&register
      summary: Registration
      operationId: registration_register_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterModel"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '409':
          description: Username already taken
          content:
            application/json:
              schema: {}
  "/login":
    post:
      tags:
      - login&register
      summary: Auth User
      operationId: auth_user_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              "$ref": "#/components/schemas/Body_auth_user_login_post"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '404':
          description: Not found
          content:
            application/json:
              schema: {}
  "/profile":
    get:
      tags:
      - user
      summary: Get Current User
      operationId: get_current_user_profile_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  "/update":
    put:
      tags:
      - user
      summary: Update
      operationId: update_update_put
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateModel"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  "/posts/new":
    post:
      tags:
      - posts
      summary: Newpost
      operationId: newpost_posts_new_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NewPostModel"
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  "/posts/delete/{post_id}":
    post:
      tags:
      - posts
      summary: Deletepost
      operationId: deletepost_posts_delete__post_id__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '204':
          description: No content
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
        
  "/posts/{post_id}":
    get:
      tags:
      - posts
      summary: Getpost
      operationId: getpost_posts__post_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
        '404':
          description: Not found
          content:
            application/json:
              schema: {}
  "/posts/update/{post_id}":
    put:
      tags:
      - posts
      summary: Updatepost
      operationId: updatepost_posts_update__post_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NewPostModel"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
        '404':
          description: Not found
          content:
            application/json:
              schema: {}
              
  "/posts":
    post:
      tags:
      - posts
      summary: Getallposts
      operationId: getallposts_posts_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AllPosts"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
components:
  schemas:
    AllPosts:
      properties:
        page_size:
          anyOf:
          - type: integer
          title: Page Size
        page_number:
          anyOf:
          - type: integer
          title: Page Number
      type: object
      required:
      - page_size
      - page_number
      title: AllPosts
    Body_auth_user_login_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: password
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_auth_user_login_post
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    NewPostModel:
      properties:
        title:
          anyOf:
          - type: string
          title: Title
        body:
          anyOf:
          - type: string
          title: Body
      type: object
      required:
      - title
      - body
      title: NewPostModel
    RegisterModel:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: RegisterModel
    UpdateModel:
      properties:
        name:
          anyOf:
          - type: string
          title: Name
        surname:
          anyOf:
          - type: string
          title: Surname
        email:
          anyOf:
          - type: string
          title: Email
        birthday:
          anyOf:
          - type: string
          title: Birthday
        phone:
          anyOf:
          - type: string
          title: Phone
      type: object
      title: UpdateModel
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: login
tags:
- name: login&register
  description: Operations with registration and authentication
- name: user
  description: Operations with user profile
- name: posts
  description: Operations with posts